feat: create an examples package
