+++ examples/uniform.py
@@ -1,8 +1,11 @@
 """A script for running a uniform agent on Tetris."""
 
+import sys
+
+sys.path.append("..")
+
 import gym
 import numpy as np
-
 from gym_simplifiedtetris.agents import UniformAgent
