+++ gym_simplifiedtetris/agents/README.md
@@ -6,7 +6,7 @@
 
 ## 1. Uniform
 
-The uniform agent selects actions uniformly at random. See [run_uniform_agent.py](https://github.com/OliverOverend/gym-simplifiedtetris/blob/master/run_uniform_agent.py) for an example of how to use it.
+The uniform agent selects actions uniformly at random. See [uniform.py](https://github.com/OliverOverend/gym-simplifiedtetris/blob/master/uniform.py) for an example of how to use it.
 
 <p align="center">
     <img src="https://github.com/OliverOverend/gym-simplifiedtetris/raw/master/assets/20x10_4.gif" width="500">
@@ -14,7 +14,7 @@ The uniform agent selects actions uniformly at random. See [run_uniform_agent.py
 
 ## 2. Q-learning
 
-Due to the curse of dimensionality, this agent struggles to learn as the grid's dimensions are increased; the size of the state-action space grows exponentially. The exploration rate parameter, epsilon, is linearly annealed over the training period. The Q-learning agent selects the action with the highest state-action value following the training period. See [run_q_learning_agent.py](https://github.com/OliverOverend/gym-simplifiedtetris/blob/master/run_q_learning_agent.py) for an example of how to use it.
+Due to the curse of dimensionality, this agent struggles to learn as the grid's dimensions are increased; the size of the state-action space grows exponentially. The exploration rate parameter, epsilon, is linearly annealed over the training period. The Q-learning agent selects the action with the highest state-action value following the training period. See [q_learning.py](https://github.com/OliverOverend/gym-simplifiedtetris/blob/master/q_learning.py) for an example of how to use it.
 
 <p align="center">
     <img src="https://github.com/OliverOverend/gym-simplifiedtetris/raw/master/assets/7x4_3_q_learning.gif" width="500">
@@ -26,7 +26,7 @@ The heuristic agent selects the action with the highest heuristic score, based o
 
 ***- landing height + eroded cells - row transitions - column transitions -4 x holes - cumulative wells***
 
-Similar to how Colin Fahey implemented Dellacherie's agent, ties are broken by selecting the action with the largest priority. Deviations from and to the left of the center of the grid are rewarded, and rotations are punished. See [run_heuristic_agent.py](https://github.com/OliverOverend/gym-simplifiedtetris/blob/master/run_heuristic_agent.py) for an example of how to use it.
+Similar to how Colin Fahey implemented Dellacherie's agent, ties are broken by selecting the action with the largest priority. Deviations from and to the left of the center of the grid are rewarded, and rotations are punished. See [heuristic.py](https://github.com/OliverOverend/gym-simplifiedtetris/blob/master/heuristic.py) for an example of how to use it.
 
 <p align="center">
     <img src="https://github.com/OliverOverend/gym-simplifiedtetris/raw/master/assets/20x10_4_heuristic.gif" width="500">
