+++ gym_simplifiedtetris/auxiliary/__init__.py
@@ -1,0 +1,4 @@
+from gym_simplifiedtetris.auxiliary.colours import Colours
+from gym_simplifiedtetris.auxiliary.polymino import Polymino
+
+__all__ = []
