+++ gym_simplifiedtetris/auxiliary/polymino.py
