+++ gym_simplifiedtetris/envs/simplified_tetris_part_binary_env.py
@@ -32,7 +32,7 @@ class SimplifiedTetrisPartBinaryEnv(_SimplifiedTetrisBaseEnv):
         return spaces.Box(
             low=low,
             high=high,
-            dtype=np.int,
+            dtype=int,
         )
 
     def _get_obs(self) -> np.array:
