--- gym_simplifiedtetris/helpers/__init__.py
@@ -1,4 +1,0 @@
-from gym_simplifiedtetris.helpers.eval_agent import eval_agent
-from gym_simplifiedtetris.helpers.train_q_learning import train_q_learning
-
-__all__ = ["eval_agent", "train_q_learning"]
