--- gym_simplifiedtetris/utils/__init__.py
@@ -1,4 +1,0 @@
-from gym_simplifiedtetris.utils.colours import Colours
-from gym_simplifiedtetris.utils.polymino import Polymino
-
-__all__ = []
