+++ requirements.txt
@@ -1,22 +1,40 @@
-build==0.7.0
+attrs==21.4.0
+black==22.3.0
+click==8.1.3
 cloudpickle==2.0.0
 cycler==0.11.0
-dataclasses==0.6
+distlib==0.3.4
+filelock==3.7.0
+fonttools==4.33.3
 gym==0.21.0
-kiwisolver==1.3.2
-matplotlib==3.4.2
-numpy==1.19.5
-opencv-python==4.5.1.48
+gym-notices==0.0.6
+importlib-metadata==4.11.3
+iniconfig==1.1.1
+kiwisolver==1.4.2
+matplotlib==3.5.2
+mypy-extensions==0.4.3
+numpy==1.22.3
+opencv-python==4.5.5.64
 packaging==21.3
-pandas==1.3.5
-pep517==0.12.0
-Pillow==8.4.0
-pyparsing==3.0.6
+pandas==1.4.2
+pathspec==0.9.0
+Pillow==9.1.0
+platformdirs==2.5.2
+pluggy==1.0.0
+py==1.11.0
+pyparsing==3.0.9
+pytest==7.1.2
+pytest-html==3.1.1
+pytest-metadata==2.0.1
 python-dateutil==2.8.2
-pytz==2021.3
+pytz==2022.1
 six==1.16.0
-stable-baselines3==1.1.0
-tomli==2.0.0
-torch==1.10.1
-tqdm==4.51.0
-typing-extensions==4.0.1
+stable-baselines3==1.5.0
+toml==0.10.2
+tomli==2.0.1
+torch==1.11.0
+tox==3.25.0
+tqdm==4.64.0
+typing_extensions==4.2.0
+virtualenv==20.14.1
+zipp==3.8.0
