--- setup.cfg
@@ -1,9 +1,0 @@
-[metadata]
-license_files = LICENSE.txt
-
-[mutmut]
-paths_to_mutate = gym_simplifiedtetris/envs
-tests = tests/
-pytest = 
-backup=False
-runner=python3 -m pytest -x
